fix: build issues
